+++ src/core/ApiConnector.ts
@@ -77,4 +77,14 @@ export class ApiConnector {
         });
         return response.data;
     }
+
+    public async toggleFluentSnippet(id: string | number, active: boolean) {
+        const response = await axios.put(`${this.apiUrl}wp-json/ide/v1/fluent-snippets/${id}/toggle`, 
+            { active: active }, 
+            {
+                headers: this.getAuthHeaders()
+            }
+        );
+        return response.data;
+    }
 }